Authenticated Controllers

+++ backend/Controllers/UserPaymentMethodController.cs
@@ -1,11 +1,13 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using backend.Data;
 using backend.Dtos.PaymentMethod;
 using backend.Interfaces;
 using backend.Mappers;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -13,6 +15,7 @@ namespace backend.Controllers
 {
     [Route("backend/paymentmethod")]
     [ApiController]
+    [Authorize]
     public class UserPaymentMethodController : ControllerBase
     {
         private readonly ApplicationDBContext _context;
@@ -23,6 +26,11 @@ namespace backend.Controllers
             _paymentMethodRepo = paymentMethodRepo;
             _context = context;
             _userRepo = userRepo;
+        }
+
+        private string GetCurrentUserEmail()
+        {
+            return User.FindFirstValue(ClaimTypes.Email); // Assuming the ID is stored in this claim
         }
 
         [HttpGet]
@@ -61,18 +69,39 @@ namespace backend.Controllers
             }
         }
 
-        [HttpPost("{uid:int}")]
-        public async Task<IActionResult> CreatePaymentMethod([FromRoute] int uid, [FromBody] CreatePaymentMethodDto paymentMethodDto)
+        // [HttpPost("{uid:int}")]
+        // public async Task<IActionResult> CreatePaymentMethod([FromRoute] int uid, [FromBody] CreatePaymentMethodDto paymentMethodDto)
+        // {
+        //     if (!ModelState.IsValid)
+        //     {
+        //         return BadRequest(ModelState);
+        //     }
+        //     if (!await _userRepo.IsUserExist(uid))
+        //     {
+        //         return BadRequest("User does not exists");
+        //     }
+        //     var paymentMethodModel = paymentMethodDto.ToUserPaymentMethodFromCreateDto(uid);
+        //     await _paymentMethodRepo.CreateAsync(paymentMethodModel);
+        //     return CreatedAtAction(nameof(GetPaymentMethodById), new { id = paymentMethodModel.UserPaymentMethodId }, paymentMethodModel.ToPaymentMethodDto());
+        // }
+        [HttpPost]
+        public async Task<IActionResult> CreatePaymentMethod([FromBody] CreatePaymentMethodDto paymentMethodDto)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
-            if (!await _userRepo.IsUserExist(uid))
+            var currUserEmail = GetCurrentUserEmail();
+            if (string.IsNullOrEmpty(currUserEmail))
             {
-                return BadRequest("User does not exists");
+                return Unauthorized($"User not found {currUserEmail}");
             }
-            var paymentMethodModel = paymentMethodDto.ToUserPaymentMethodFromCreateDto(uid);
+            var currUser = await _context.Users.FirstOrDefaultAsync(u => u.UserEmail == currUserEmail);
+            if (currUser == null)
+            {
+                return NotFound($"User details not found {currUserEmail}");
+            }
+            var paymentMethodModel = paymentMethodDto.ToUserPaymentMethodFromCreateDto(currUser.UserId);
             await _paymentMethodRepo.CreateAsync(paymentMethodModel);
             return CreatedAtAction(nameof(GetPaymentMethodById), new { id = paymentMethodModel.UserPaymentMethodId }, paymentMethodModel.ToPaymentMethodDto());
         }
@@ -84,6 +113,21 @@ namespace backend.Controllers
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
+            }
+            var currUserEmail = GetCurrentUserEmail();
+            if (string.IsNullOrEmpty(currUserEmail))
+            {
+                return Unauthorized($"User not found {currUserEmail}");
+            }
+            var currUser = await _context.Users.FirstOrDefaultAsync(u => u.UserEmail == currUserEmail);
+            if (currUser == null)
+            {
+                return NotFound($"User details not found {currUserEmail}");
+            }
+            var updatingPaymentMethodModel = await _context.UserPaymentMethods.FirstOrDefaultAsync(pm => pm.UserPaymentMethodId == id);
+            if (updatingPaymentMethodModel.UserId != currUser.UserId)
+            {
+                return Unauthorized($"User has no access to update");
             }
             var paymentMethodModel = await _paymentMethodRepo.UpdateAsync(id, paymentMethodDto);
             if (paymentMethodModel == null)
@@ -97,6 +141,21 @@ namespace backend.Controllers
         [Route("{id:int}")]
         public async Task<IActionResult> DeletePaymentMethod([FromRoute] int id)
         {
+            var currUserEmail = GetCurrentUserEmail();
+            if (string.IsNullOrEmpty(currUserEmail))
+            {
+                return Unauthorized($"User not found {currUserEmail}");
+            }
+            var currUser = await _context.Users.FirstOrDefaultAsync(u => u.UserEmail == currUserEmail);
+            if (currUser == null)
+            {
+                return NotFound($"User details not found {currUserEmail}");
+            }
+            var deletingPaymentMethodModel = await _context.UserPaymentMethods.FirstOrDefaultAsync(pm => pm.UserPaymentMethodId == id);
+            if (deletingPaymentMethodModel.UserId != currUser.UserId)
+            {
+                return Unauthorized($"User has no access to delete");
+            }
             var paymentMethodModel = await _paymentMethodRepo.DeleteAsync(id);
             if (paymentMethodModel == null)
             {
